split out mapsnap/keymap subdir
